Add a temporary solution for modal dialog centered alignment
Spec: https://html.spec.whatwg.org/multipage/#centered-alignment

The spec is opaque about the centered alignment mode and we couldn't
come up a satisfied solution by using the current version of the
spec. There will be discussion in CSSWG to talk about better solutions.
In the meantime, we are adding this patch as a temporary solution with
basic centered alignment support to allow us enabling the feature
in Nightly for more testing.

Differential Revision: https://phabricator.services.mozilla.com/D77678

bugzilla-url: https://bugzilla.mozilla.org/show_bug.cgi?id=1642364
gecko-commit: 3d06403ffd5cea9cd14680acb434bfe3b85cbde2
gecko-integration-branch: autoland
gecko-reviewers: emilio