Move non autogenerated canvas wpt tests into manual folder
It's not altogether obvious what the outcome of
wpt/html/canvas/tools/build.sh is meant to be as there are a few dozen
non-autogenerated tests throughout the repo. Here all tests that are
not the result of build.sh are put into a "manual" folder.

Next I'll do it for offscreen tests, but this CL is big enough as is.

Bug: 1091267
Change-Id: I7971558b046ce98dacd7fbcb4a4c5ceba84ba171