Improve background position snapping
A previous fix to ensure painting backgrounds hard up against the
edge of the positioning area has led to images at other positions
dancing around due to sub-pixel snapping.

Change the way we snap the position, again, in an attempt to improve
the situation. This does not totally fix the animated transition case
because that has an infinite (sort of) set of sub-pixel sizes and not
all of them will snap the same way. The latter seems nearly impossible
to fix.

Bug: 1088406
Change-Id: I5549404378e7668c8edcecb9816ba4d401bcdde0